more try and error in plots and renamed one file
